INITIAL SETUP - Addresses issue on cold install / adds more instructions and a Makefile convenience function to do setup "make web-setup"
From Yvette:
Frontend and backend setup
When running the code via "make web" this will fail if following the primary readme instructions because Vite and Uvicorn are not found. The setup of the college_board_eval/web/backend and college_board_eval/web/frontend have to be done first.

Action: Instructions in the primary README would suffice. Also, could create a Makefile block that runs the setup of both as a convenience.